Move storage into plugins
