--- src/common/store/index.ts
@@ -1,2 +1,0 @@
-export { Store } from './Store';
-export { TrieStore } from './trieStore';
