+++ src/index.ts
@@ -2,12 +2,12 @@ import { ConsoleLogger } from './plugins/loggers';
 import { DNSServer } from './server';
 import { DNSOverTCP, DNSOverUDP } from './common/network';
 import { Handler } from './server';
-import { TrieStore } from './common/store';
+import { DefaultStore } from './plugins/storage';
 
 export * as network from './common/network';
 export * as logging from './plugins/loggers';
 export * as server from './server';
-export * as store from './common/store';
+export * as store from './plugins/storage';
 
 export {
   DNSOverTCP,
@@ -21,7 +21,7 @@ export {
 } from './common/network';
 export { ConsoleLogger, Logger } from './plugins/loggers';
 export { DNSServer, DNSRequest, DNSResponse, Handler, NextFunction } from './server';
-export { TrieStore, Store } from './common/store';
+export { DefaultStore, Store } from './plugins/storage';
 export { DefaultRouter, Router } from './common/router';
 
 const logger = new ConsoleLogger(true, true);
@@ -32,7 +32,7 @@ const s = new DNSServer({
   },
 });
 
-const store = new TrieStore();
+const store = new DefaultStore();
 store.set('example.com', 'A', {
   name: 'example.com',
   type: 'A',
