+++ src/plugins/storage/index.ts
@@ -1,0 +1,2 @@
+export { Store } from './Store';
+export { DefaultStore } from './DefaultStore';
